pyproject.toml
[project]
name = "pharmacy-stock-ocr"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.6",
    "mistralai>=1.0.0",
    "openai>=1.12.0",
    "google-auth>=2.27.0",
    "google-api-python-client>=2.116.0",
    "google-cloud-storage>=2.14.0",
    "pillow>=10.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
]
